running extremely long load time, moving to branch for help w debugging
